refactor: 重構模組管理相關代碼，移除 runtime_typing 並整合至 online_mod_queue，更新相關引用
